Module language_common::error_codes

Constants

EASSERT_ERROR

EBAD_ACCOUNT_AUTHENTICATION_KEY

EBAD_SIGNATURE
Error codes that can be emitted by the prologue. These have special significance to the VM when they are raised during the prologue. However, they can also be raised by user code during execution of a transaction script. They have no significance to the VM in that case.

ECANT_PAY_GAS_DEPOSIT

EGAS_DEPOSIT_TOO_SMALL

EINSUFFICIENT_BALANCE
Generic error codes. These codes don't have any special meaning for the VM, but they are useful conventions for debugging

ESEQUENCE_NUMBER_TOO_NEW

ESEQUENCE_NUMBER_TOO_OLD