Trait language_common::storage_adapter::cache::GlobalStorageCache

pub trait GlobalStorageCache {
    fn read_value(
        &mut self, 
        key: &ResourceKey, 
        accesses: &Accesses
    ) -> Option<Value>;

    fn read_code(&mut self, key: &CodeKey) -> Option<&[u8]>;

    fn write_code(&mut self, key: &CodeKey, code: Vec<u8>);

    fn write_value(
        &mut self, 
        key: &ResourceKey, 
        accesses: &Accesses, 
        value: Value
    );

    fn delete_value(&mut self, key: &ResourceKey, accesses: &Accesses);

    fn write_set(&self) -> WriteSet;

    fn clear_cache(&mut self);

    fn modified_addresses(&self) -> BTreeSet<AccountAddress>;
}

Required methods

fn read_value(
&mut self,
key: &ResourceKey,
accesses: &Accesses
) -> Option<Value>

First look up the value associated with address and accesses in the local cache. If it wasn't in the cache, fetch the version that has been stored globally and add it to the cache. Fails if key is unbound.

fn read_code(&mut self, key: &CodeKey) -> Option<&[u8]>

Look up the struct associated with key in the global store

fn write_code(&mut self, key: &CodeKey, code: Vec<u8>)

Publish struct class with key under key.address.

fn write_value(&mut self, key: &ResourceKey, accesses: &Accesses, value: Value)

Write a value temporarily to address and accesses in the local cache. Change should be immediately visible and won't be flushed to the storagedb.

fn delete_value(&mut self, key: &ResourceKey, accesses: &Accesses)

Write a tombstone value to the access path key.accesses and all of its suffixes. Panics if there is no value at key.accesses.

fn write_set(&self) -> WriteSet

Generate a mapping that contains all proposed write to the storagedb due to the current execution.

fn clear_cache(&mut self)

Flush the entire cache out to a clean state.

fn modified_addresses(&self) -> BTreeSet<AccountAddress>

Return the set of account addresses whose values have been modified during the current execution

Loading content...